fix(client): extract iframe preflight utils to fix vitest resolution
Extract shouldUseIframePreflightProxy and shouldUseServerIframePreflight
from useIframePreflight.ts into iframePreflight.utils.ts so the test can
import pure functions without pulling in React hooks and @/ aliased
modules that fail under vitest workspace resolution.
